+++ libs/xkontiVector2d.h
@@ -12,15 +12,15 @@
 // GLOBAL CONSTANTS
 //////////////////////////////////////////
 
-const float PI = 3.141592653589793238462643383279;
+const double PI = 3.141592653589793238462643383279;
 
 
 //////////////////////////////////////////
 // MISC FUNCTIONS
 //////////////////////////////////////////
 
-float radToDeg(float _rad);		// Convert radians to degrees
-float degToRad(float _deg);		// Convert degrees to radians
+double radToDeg(double _rad);		// Convert radians to degrees
+double degToRad(double _deg);		// Convert degrees to radians
 
 
 //////////////////////////////////////////
@@ -31,8 +31,8 @@ class Vector2D {
 public:
 	// Constructor & Destructor
 	Vector2D();
-	Vector2D(float _rad);
-	Vector2D(float _x, float _y);
+	Vector2D(double _rad);
+	Vector2D(double _x, double _y);
 	~Vector2D();
 
 	// Operators
@@ -40,26 +40,26 @@ public:
 	Vector2D operator + (Vector2D _a);
 	Vector2D operator - (Vector2D _a);
 	Vector2D operator * (Vector2D _a);
-	Vector2D operator * (float _a);
+	Vector2D operator * (double _a);
 	Vector2D operator / (Vector2D _a);
-	Vector2D operator / (float _a);
+	Vector2D operator / (double _a);
 
 	// Constants
 	inline Vector2D forward();
 
 	// Functions
-	float rad();		// Returns direction of vector in radians
-	float length();	// Returns length of vector
+	double rad();		// Returns direction of vector in radians
+	double length();	// Returns length of vector
 	Vector2D normalized();			// Returns normalized vector
-	Vector2D rotated(float _rad);	// Returns rotated vector
+	Vector2D rotated(double _rad);	// Returns rotated vector
 
 	// Procedures
 	void normalize();		// Normalizes vector
-	void rotate(float _rad);	// Rotate vector
+	void rotate(double _rad);	// Rotate vector
 
 	// Contents
-	float x;
-	float y;
+	double x;
+	double y;
 private:
 
 };
